Fix #94

+++ src/main/java/com/klikli_dev/modonomicon/client/gui/book/BookContentScreen.java
@@ -596,10 +596,14 @@ public class BookContentScreen extends Screen implements BookScreenWithButtons{
     public List<Component> getTooltipFromItem(ItemStack pItemStack) {
         var tooltip = super.getTooltipFromItem(pItemStack);
 
-        if(this.isHoveringItemLink && ModonomiconJeiIntegration.isJeiLoaded()){
+        if(this.isHoveringItemLink){
             tooltip.add(new TextComponent(""));
-            tooltip.add(new TranslatableComponent(Gui.HOVER_ITEM_LINK_INFO).withStyle(Style.EMPTY.withItalic(true).withColor(ChatFormatting.GREEN)));
-            tooltip.add(new TranslatableComponent(Gui.HOVER_ITEM_LINK_INFO_LINE2).withStyle(Style.EMPTY.withItalic(true).withColor(ChatFormatting.GRAY)));
+            if(ModonomiconJeiIntegration.isJeiLoaded()){
+                tooltip.add(new TranslatableComponent(Gui.HOVER_ITEM_LINK_INFO).withStyle(Style.EMPTY.withItalic(true).withColor(ChatFormatting.GREEN)));
+                tooltip.add(new TranslatableComponent(Gui.HOVER_ITEM_LINK_INFO_LINE2).withStyle(Style.EMPTY.withItalic(true).withColor(ChatFormatting.GRAY)));
+            } else {
+                tooltip.add(new TranslatableComponent(Gui.HOVER_ITEM_LINK_INFO_NO_JEI).withStyle(Style.EMPTY.withItalic(true).withColor(ChatFormatting.RED)));
+            }
         }
 
         return tooltip;
@@ -656,27 +660,27 @@ public class BookContentScreen extends Screen implements BookScreenWithButtons{
                     }
 
                     if(ItemLinkRenderer.isItemLink(event.getValue())){
+                        if( ModonomiconJeiIntegration.isJeiLoaded()){
+                            var itemId = event.getValue().substring(ItemLinkRenderer.PROTOCOL_ITEM_LENGTH);
+                            var itemStack = ItemStackUtil.loadFromParsed(ItemStackUtil.parseItemStackString(itemId));
 
-                        var itemId = event.getValue().substring(ItemLinkRenderer.PROTOCOL_ITEM_LENGTH);
-                        var itemStack = ItemStackUtil.loadFromParsed(ItemStackUtil.parseItemStackString(itemId));
+                            this.onClose(); //we have to do this before showing JEI, because super.onClose() clears Gui Layers, and thus would kill JEIs freshly spawned gui
 
-                        this.onClose(); //we have to do this before showing JEI, because super.onClose() clears Gui Layers, and thus would kill JEIs freshly spawned gui
+                            if(Screen.hasShiftDown()){
+                                ModonomiconJeiIntegration.showUses(itemStack);
+                            } else {
+                                ModonomiconJeiIntegration.showRecipe(itemStack);
+                            }
 
-                        if(Screen.hasShiftDown()){
-                            ModonomiconJeiIntegration.showUses(itemStack);
-                        } else {
-                            ModonomiconJeiIntegration.showRecipe(itemStack);
+                            if(!ModonomiconJeiIntegration.isJEIRecipesGuiOpen()){
+                                this.minecraft.pushGuiLayer(this);
+                            }
+
+                            //TODO: Consider adding logic to restore content screen after JEI gui close
+                            //      currently only the overview screen is restored (because JEI does not use Forges Gui Stack, only vanilla screen, thus only saves one parent screen)
+                            //      we could fix that by listening to the Closing event from forge, and in that set the closing time
+                            //      -> then on init of overview screen, if closing time is < delta, push last content screen from gui manager
                         }
-
-                        if(!(this.minecraft.screen instanceof IRecipesGui)){
-                            //if we did not open a JEI gui, restore self
-                            this.minecraft.pushGuiLayer(this);
-                        }
-
-                        //TODO: Consider adding logic to restore content screen after JEI gui close
-                        //      currently only the overview screen is restored (because JEI does not use Forges Gui Stack, only vanilla screen, thus only saves one parent screen)
-                        //      we could fix that by listening to the Closing event from forge, and in that set the closing time
-                        //      -> then on init of overview screen, if closing time is < delta, push last content screen from gui manager
 
                         return true;
                     }
